% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_monitor_clarify_model_monitoring.R
\name{ClarifyBaseliningJob}
\alias{ClarifyBaseliningJob}
\title{ClarifyBaseliningJob class}
\description{
Provides functionality to retrieve baseline-specific output from Clarify baselining job.
}
\section{Super classes}{
\code{\link[sagemaker.common:.Job]{sagemaker.common::.Job}} -> \code{\link[sagemaker.common:ProcessingJob]{sagemaker.common::ProcessingJob}} -> \code{\link[sagemaker.mlcore:BaseliningJob]{sagemaker.mlcore::BaseliningJob}} -> \code{ClarifyBaseliningJob}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ClarifyBaseliningJob$new()}}
\item \href{#method-baseline_statistics}{\code{ClarifyBaseliningJob$baseline_statistics()}}
\item \href{#method-suggested_constraints}{\code{ClarifyBaseliningJob$suggested_constraints()}}
\item \href{#method-clone}{\code{ClarifyBaseliningJob$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic=".Job" data-id="format">}\href{../../sagemaker.common/html/.Job.html#method-format}{\code{sagemaker.common::.Job$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ProcessingJob" data-id="describe">}\href{../../sagemaker.common/html/ProcessingJob.html#method-describe}{\code{sagemaker.common::ProcessingJob$describe()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ProcessingJob" data-id="from_processing_arn">}\href{../../sagemaker.common/html/ProcessingJob.html#method-from_processing_arn}{\code{sagemaker.common::ProcessingJob$from_processing_arn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ProcessingJob" data-id="from_processing_name">}\href{../../sagemaker.common/html/ProcessingJob.html#method-from_processing_name}{\code{sagemaker.common::ProcessingJob$from_processing_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ProcessingJob" data-id="prepare_app_specification">}\href{../../sagemaker.common/html/ProcessingJob.html#method-prepare_app_specification}{\code{sagemaker.common::ProcessingJob$prepare_app_specification()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ProcessingJob" data-id="prepare_output_config">}\href{../../sagemaker.common/html/ProcessingJob.html#method-prepare_output_config}{\code{sagemaker.common::ProcessingJob$prepare_output_config()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ProcessingJob" data-id="prepare_processing_resources">}\href{../../sagemaker.common/html/ProcessingJob.html#method-prepare_processing_resources}{\code{sagemaker.common::ProcessingJob$prepare_processing_resources()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ProcessingJob" data-id="prepare_stopping_condition">}\href{../../sagemaker.common/html/ProcessingJob.html#method-prepare_stopping_condition}{\code{sagemaker.common::ProcessingJob$prepare_stopping_condition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ProcessingJob" data-id="start_new">}\href{../../sagemaker.common/html/ProcessingJob.html#method-start_new}{\code{sagemaker.common::ProcessingJob$start_new()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ProcessingJob" data-id="stop">}\href{../../sagemaker.common/html/ProcessingJob.html#method-stop}{\code{sagemaker.common::ProcessingJob$stop()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ProcessingJob" data-id="wait">}\href{../../sagemaker.common/html/ProcessingJob.html#method-wait}{\code{sagemaker.common::ProcessingJob$wait()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="BaseliningJob" data-id="from_processing_job">}\href{../../sagemaker.mlcore/html/BaseliningJob.html#method-from_processing_job}{\code{sagemaker.mlcore::BaseliningJob$from_processing_job()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a ClarifyBaseliningJob that tracks a baselining job by suggest_baseline()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningJob$new(processing_job)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{processing_job}}{(sagemaker.processing.ProcessingJob): The ProcessingJob used for
baselining instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baseline_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-baseline_statistics}{}}}
\subsection{Method \code{baseline_statistics()}}{
Not implemented.
             The class doesn't support statistics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningJob$baseline_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-suggested_constraints"></a>}}
\if{latex}{\out{\hypertarget{method-suggested_constraints}{}}}
\subsection{Method \code{suggested_constraints()}}{
Returns a sagemaker.model_monitor.
             Constraints object representing the constraints JSON file generated by this baselining job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningJob$suggested_constraints(file_name = NULL, kms_key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{(str): Keep this parameter to align with method signature in super class,
but it will be ignored.}

\item{\code{kms_key}}{(str): The kms key to use when retrieving the file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sagemaker.model_monitor.Constraints: The Constraints object representing the file that
             was generated by the job.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClarifyBaseliningJob$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
