% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner.R
\name{create_identical_dataset_and_algorithm_tuner}
\alias{create_identical_dataset_and_algorithm_tuner}
\title{Creates a new tuner with an identical dataset and algorithm.}
\usage{
create_identical_dataset_and_algorithm_tuner(
  parent,
  additional_parents = NULL,
  sagemaker_session = NULL
)
}
\arguments{
\item{parent}{(str): Primary parent tuning job's name from which the Tuner and
Estimator configuration has to be copied}

\item{additional_parents}{(set{str}): Set of additional parent tuning job's
names along with the primary parent tuning job name to be used in
warm starting the transfer learning tuner.}

\item{sagemaker_session}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other AWS
services needed. If not specified, one is created using the default
AWS configuration chain.}
}
\value{
sagemaker.tuner.HyperparameterTuner: a new ``HyperparameterTuner``
             object for the warm-started hyperparameter tuning job
}
\description{
It does this identical creation by copying the request fields from the
             provided parent to the new instance of ``HyperparameterTuner`` followed
             by addition of warm start configuration with the type as
             "IdenticalDataAndAlgorithm" and ``parents`` as the union of provided list
             of ``additional_parents`` and the ``parent``.
}
