% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner.R
\name{create_transfer_learning_tuner}
\alias{create_transfer_learning_tuner}
\title{Creates a new ``HyperParameterTuner`` instance from the parent.}
\usage{
create_transfer_learning_tuner(
  parent,
  additional_parents = NULL,
  estimator = NULL,
  sagemaker_session = NULL
)
}
\arguments{
\item{parent}{(str): Primary parent tuning job's name from which the Tuner and
Estimator configuration has to be copied}

\item{additional_parents}{(set{str}): Set of additional parent tuning job's
names along with the primary parent tuning job name to be used in
warm starting the identical dataset and algorithm tuner.}

\item{estimator}{(sagemaker.estimator.EstimatorBase): An estimator object that
has been initialized with the desired configuration. There does not
need to be a training job associated with this instance.}

\item{sagemaker_session}{(sagemaker.session.Session): Session object which
manages interactions with Amazon SageMaker APIs and any other AWS
services needed. If not specified, one is created using the default
AWS configuration chain.}
}
\value{
sagemaker.tuner.HyperparameterTuner: New instance of warm started
             HyperparameterTuner
}
\description{
It creates the new tuner by copying the request fields from the provided
             parent to the new instance of ``HyperparameterTuner`` followed by addition
             of warm start configuration with the type as "TransferLearning" and
             ``parents`` as the union of provided list of ``additional_parents`` and
             the ``parent``.
}
