% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_hyperparameter.R
\name{DataTypes}
\alias{DataTypes}
\title{Helper class to check data types}
\description{
Throw error when failed to convert object
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-float}{\code{DataTypes$float()}}
\item \href{#method-str}{\code{DataTypes$str()}}
\item \href{#method-int}{\code{DataTypes$int()}}
\item \href{#method-bool}{\code{DataTypes$bool()}}
\item \href{#method-clone}{\code{DataTypes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-float"></a>}}
\if{latex}{\out{\hypertarget{method-float}{}}}
\subsection{Method \code{float()}}{
set numeric type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypes$float(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(object): R object to be converted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-str"></a>}}
\if{latex}{\out{\hypertarget{method-str}{}}}
\subsection{Method \code{str()}}{
set character type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypes$str(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(object): R object to be converted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-int"></a>}}
\if{latex}{\out{\hypertarget{method-int}{}}}
\subsection{Method \code{int()}}{
set integer type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypes$int(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(object): R object to be converted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bool"></a>}}
\if{latex}{\out{\hypertarget{method-bool}{}}}
\subsection{Method \code{bool()}}{
set boolean type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypes$bool(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(object): R object to be converted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataTypes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
