% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_ipinsights.R
\name{IPInsightsPredictor}
\alias{IPInsightsPredictor}
\title{Returns dot product of entity and IP address embeddings as a score for
             compatibility.}
\description{
The implementation of
             :meth:`~sagemaker.predictor.Predictor.predict` in this
             `Predictor` requires a numpy ``ndarray`` as input. The array should
             contain two columns. The first column should contain the entity ID. The
             second column should contain the IPv4 address in dot notation.
}
\section{Super classes}{
\code{\link[sagemaker.mlcore:PredictorBase]{sagemaker.mlcore::PredictorBase}} -> \code{\link[sagemaker.mlcore:Predictor]{sagemaker.mlcore::Predictor}} -> \code{IPInsightsPredictor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IPInsightsPredictor$new()}}
\item \href{#method-clone}{\code{IPInsightsPredictor$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="delete_endpoint">}\href{../../sagemaker.mlcore/html/Predictor.html#method-delete_endpoint}{\code{sagemaker.mlcore::Predictor$delete_endpoint()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="delete_model">}\href{../../sagemaker.mlcore/html/Predictor.html#method-delete_model}{\code{sagemaker.mlcore::Predictor$delete_model()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="delete_predictor">}\href{../../sagemaker.mlcore/html/Predictor.html#method-delete_predictor}{\code{sagemaker.mlcore::Predictor$delete_predictor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="disable_data_capture">}\href{../../sagemaker.mlcore/html/Predictor.html#method-disable_data_capture}{\code{sagemaker.mlcore::Predictor$disable_data_capture()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="enable_data_capture">}\href{../../sagemaker.mlcore/html/Predictor.html#method-enable_data_capture}{\code{sagemaker.mlcore::Predictor$enable_data_capture()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="endpoint_context">}\href{../../sagemaker.mlcore/html/Predictor.html#method-endpoint_context}{\code{sagemaker.mlcore::Predictor$endpoint_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="format">}\href{../../sagemaker.mlcore/html/Predictor.html#method-format}{\code{sagemaker.mlcore::Predictor$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="list_monitors">}\href{../../sagemaker.mlcore/html/Predictor.html#method-list_monitors}{\code{sagemaker.mlcore::Predictor$list_monitors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="predict">}\href{../../sagemaker.mlcore/html/Predictor.html#method-predict}{\code{sagemaker.mlcore::Predictor$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="update_data_capture_config">}\href{../../sagemaker.mlcore/html/Predictor.html#method-update_data_capture_config}{\code{sagemaker.mlcore::Predictor$update_data_capture_config()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="update_endpoint">}\href{../../sagemaker.mlcore/html/Predictor.html#method-update_endpoint}{\code{sagemaker.mlcore::Predictor$update_endpoint()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize IPInsightsPredictor class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IPInsightsPredictor$new(endpoint_name, sagemaker_session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_name}}{(str): Name of the Amazon SageMaker endpoint to which
requests are sent.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): A SageMaker Session
object, used for SageMaker interactions (default: None). If not
specified, one is created using the default AWS configuration
chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IPInsightsPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
