% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_model.R
\name{LambdaModel}
\alias{LambdaModel}
\title{LambdaModel class}
\description{
A model that can be deployed to Lambda
}
\section{Super class}{
\code{\link[sagemaker.common:ModelBase]{sagemaker.common::ModelBase}} -> \code{LambdaModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LambdaModel$new()}}
\item \href{#method-deploy}{\code{LambdaModel$deploy()}}
\item \href{#method-delete_model}{\code{LambdaModel$delete_model()}}
\item \href{#method-clone}{\code{LambdaModel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ModelBase" data-id="format">}\href{../../sagemaker.common/html/ModelBase.html#method-format}{\code{sagemaker.common::ModelBase$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.common" data-topic="ModelBase" data-id="help">}\href{../../sagemaker.common/html/ModelBase.html#method-help}{\code{sagemaker.common::ModelBase$help()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize instance attributes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaModel$new(image_uri, role, client = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image_uri}}{: URI of a container image in the Amazon ECR registry. The image
should contain a handler that performs inference.}

\item{\code{role}}{: The Amazon Resource Name (ARN) of the IAM role that Lambda will assume
when it performs inference}

\item{\code{client}}{: The Lambda client used to interact with Lambda.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deploy"></a>}}
\if{latex}{\out{\hypertarget{method-deploy}{}}}
\subsection{Method \code{deploy()}}{
Create a Lambda function using the image specified in the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaModel$deploy(function_name, timeout, memory_size, wait = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{function_name}}{: The name of the function.}

\item{\code{timeout}}{: The number of seconds that the function can run for before being terminated.}

\item{\code{memory_size}}{: The amount of memory in MB that the function has access to.}

\item{\code{wait}}{: If true, wait until the deployment completes (default: True).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A LambdaPredictor instance that performs inference using the specified image.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_model"></a>}}
\if{latex}{\out{\hypertarget{method-delete_model}{}}}
\subsection{Method \code{delete_model()}}{
Destroy resources associated with this model.
             This method does not delete the image specified in the constructor. As
             a result, this method is a no-op.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaModel$delete_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
