% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverless_predictor.R
\name{LambdaPredictor}
\alias{LambdaPredictor}
\title{LamdbaPredictor}
\description{
A deployed model hosted on Lambda.
}
\section{Super class}{
\code{\link[sagemaker.mlcore:PredictorBase]{sagemaker.mlcore::PredictorBase}} -> \code{LambdaPredictor}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{content_type}}{The MIME type of the data sent to the Lambda function.}

\item{\code{accept}}{The content type(s) that are expected from the Lambda function.}

\item{\code{function_name}}{The name of the Lambda function this predictor invokes.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LambdaPredictor$new()}}
\item \href{#method-predict}{\code{LambdaPredictor$predict()}}
\item \href{#method-delete_predictor}{\code{LambdaPredictor$delete_predictor()}}
\item \href{#method-clone}{\code{LambdaPredictor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize instance attributes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaPredictor$new(function_name, client = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{function_name}}{: The name of the function.}

\item{\code{client}}{: The Lambda client used to interact with Lambda.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Invoke the Lambda function specified in the constructor.
             This function is synchronous. It will only return after the function
             has produced a prediction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaPredictor$predict(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{: The data sent to the Lambda function as input.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The data returned by the Lambda function.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_predictor"></a>}}
\if{latex}{\out{\hypertarget{method-delete_predictor}{}}}
\subsection{Method \code{delete_predictor()}}{
Destroy the Lambda function specified in the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaPredictor$delete_predictor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LambdaPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
