% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_linear_learner.R
\name{LinearLearnerPredictor}
\alias{LinearLearnerPredictor}
\title{Performs binary-classification or regression prediction from input vectors.}
\description{
The implementation of
             :meth:`~sagemaker.predictor.Predictor.predict` in this
             `Predictor` requires a numpy ``ndarray`` as input. The array should
             contain the same number of columns as the feature-dimension of the data used
             to fit the model this Predictor performs inference on.
             :func:`predict` returns a list of
             :class:`~sagemaker.amazon.record_pb2.Record` objects, one for each row in
             the input ``ndarray``. The prediction is stored in the ``"predicted_label"``
             key of the ``Record.label`` field.
}
\section{Super classes}{
\code{\link[sagemaker.mlcore:PredictorBase]{sagemaker.mlcore::PredictorBase}} -> \code{\link[sagemaker.mlcore:Predictor]{sagemaker.mlcore::Predictor}} -> \code{LinearLearnerPredictor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LinearLearnerPredictor$new()}}
\item \href{#method-clone}{\code{LinearLearnerPredictor$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="delete_endpoint">}\href{../../sagemaker.mlcore/html/Predictor.html#method-delete_endpoint}{\code{sagemaker.mlcore::Predictor$delete_endpoint()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="delete_model">}\href{../../sagemaker.mlcore/html/Predictor.html#method-delete_model}{\code{sagemaker.mlcore::Predictor$delete_model()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="delete_predictor">}\href{../../sagemaker.mlcore/html/Predictor.html#method-delete_predictor}{\code{sagemaker.mlcore::Predictor$delete_predictor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="disable_data_capture">}\href{../../sagemaker.mlcore/html/Predictor.html#method-disable_data_capture}{\code{sagemaker.mlcore::Predictor$disable_data_capture()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="enable_data_capture">}\href{../../sagemaker.mlcore/html/Predictor.html#method-enable_data_capture}{\code{sagemaker.mlcore::Predictor$enable_data_capture()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="endpoint_context">}\href{../../sagemaker.mlcore/html/Predictor.html#method-endpoint_context}{\code{sagemaker.mlcore::Predictor$endpoint_context()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="format">}\href{../../sagemaker.mlcore/html/Predictor.html#method-format}{\code{sagemaker.mlcore::Predictor$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="list_monitors">}\href{../../sagemaker.mlcore/html/Predictor.html#method-list_monitors}{\code{sagemaker.mlcore::Predictor$list_monitors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="predict">}\href{../../sagemaker.mlcore/html/Predictor.html#method-predict}{\code{sagemaker.mlcore::Predictor$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="update_data_capture_config">}\href{../../sagemaker.mlcore/html/Predictor.html#method-update_data_capture_config}{\code{sagemaker.mlcore::Predictor$update_data_capture_config()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Predictor" data-id="update_endpoint">}\href{../../sagemaker.mlcore/html/Predictor.html#method-update_endpoint}{\code{sagemaker.mlcore::Predictor$update_endpoint()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize LinearLearnerPredictor Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinearLearnerPredictor$new(endpoint_name, sagemaker_session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint_name}}{(str): Name of the Amazon SageMaker endpoint to which
requests are sent.}

\item{\code{sagemaker_session}}{(sagemaker.session.Session): A SageMaker Session
object, used for SageMaker interactions (default: None). If not
specified, one is created using the default AWS configuration
chain.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinearLearnerPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
