% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{model_parameters}
\alias{model_parameters}
\title{An array of parameters for modelling methods}
\format{
An object of class \code{Enum} (inherits from \code{environment}) of length 8.
}
\usage{
model_parameters
}
\description{
An array of parameters for modelling methods
}
\keyword{internal}
