% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ModelBase}
\alias{ModelBase}
\title{An object that encapsulates a trained model.}
\description{
Models can be deployed to compute services like a SageMaker ``Endpoint``
       or Lambda. Deployed models can be used to perform real-time inference.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-deploy}{\code{ModelBase$deploy()}}
\item \href{#method-delete_model}{\code{ModelBase$delete_model()}}
\item \href{#method-format}{\code{ModelBase$format()}}
\item \href{#method-help}{\code{ModelBase$help()}}
\item \href{#method-clone}{\code{ModelBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deploy"></a>}}
\if{latex}{\out{\hypertarget{method-deploy}{}}}
\subsection{Method \code{deploy()}}{
Deploy this model to a compute service.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBase$deploy(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{: not currently implemented}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_model"></a>}}
\if{latex}{\out{\hypertarget{method-delete_model}{}}}
\subsection{Method \code{delete_model()}}{
Destroy resources associated with this model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBase$delete_model(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{: not currently implemented}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBase$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-help"></a>}}
\if{latex}{\out{\hypertarget{method-help}{}}}
\subsection{Method \code{help()}}{
Return class documentation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBase$help()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
