% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarify.R
\name{ModelConfig}
\alias{ModelConfig}
\title{Model Config}
\description{
Config object related to a model and its endpoint to be created.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{predictor_config}}{Predictor dictionary of the analysis config}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelConfig$new()}}
\item \href{#method-get_predictor_config}{\code{ModelConfig$get_predictor_config()}}
\item \href{#method-format}{\code{ModelConfig$format()}}
\item \href{#method-clone}{\code{ModelConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a configuration of a model and the endpoint to be created for it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfig$new(
  model_name,
  instance_count,
  instance_type,
  accept_type = NULL,
  content_type = NULL,
  content_template = NULL,
  custom_attributes = NULL,
  accelerator_type = NULL,
  endpoint_name_prefix = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{(str): Model name (as created by 'CreateModel').}

\item{\code{instance_count}}{(int): The number of instances of a new endpoint for model inference.}

\item{\code{instance_type}}{(str): The type of EC2 instance to use for model inference,
for example, 'ml.c5.xlarge'.}

\item{\code{accept_type}}{(str): The model output format to be used for getting inferences with the
shadow endpoint. Valid values are "text/csv" for CSV and "application/jsonlines".
Default is the same as content_type.}

\item{\code{content_type}}{(str): The model input format to be used for getting inferences with the
shadow endpoint. Valid values are "text/csv" for CSV and "application/jsonlines".
Default is the same as dataset format.}

\item{\code{content_template}}{(str): A template string to be used to construct the model input from
dataset instances. It is only used when "model_content_type" is
"application/jsonlines". The template should have one and only one placeholder
$features which will be replaced by a features list for to form the model inference
input.}

\item{\code{custom_attributes}}{(str): Provides additional information about a request for an
inference submitted to a model hosted at an Amazon SageMaker endpoint. The
information is an opaque value that is forwarded verbatim. You could use this
value, for example, to provide an ID that you can use to track a request or to
provide other metadata that a service endpoint was programmed to process. The value
must consist of no more than 1024 visible US-ASCII characters as specified in
Section 3.3.6. Field Value Components (
\url{https://tools.ietf.org/html/rfc7230#section-3.2.6}) of the Hypertext Transfer
Protocol (HTTP/1.1).}

\item{\code{accelerator_type}}{(str): The Elastic Inference accelerator type to deploy to the model
endpoint instance for making inferences to the model, see
\url{https://docs.aws.amazon.com/sagemaker/latest/dg/ei.html}.}

\item{\code{endpoint_name_prefix}}{(str): The endpoint name prefix of a new endpoint. Must follow
pattern "^[a-zA-Z0-9](-\*[a-zA-Z0-9]".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_predictor_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_predictor_config}{}}}
\subsection{Method \code{get_predictor_config()}}{
Returns part of the predictor dictionary of the analysis config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfig$get_predictor_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfig$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
