% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_metrics.R
\name{ModelMetrics}
\alias{ModelMetrics}
\title{ModelMetrics class}
\description{
Accepts model metrics parameters for conversion to request dict.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelMetrics$new()}}
\item \href{#method-to_request_list}{\code{ModelMetrics$to_request_list()}}
\item \href{#method-format}{\code{ModelMetrics$format()}}
\item \href{#method-clone}{\code{ModelMetrics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``ModelMetrics`` instance and turn parameters into dict.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMetrics$new(
  model_statistics = NULL,
  model_constraints = NULL,
  model_data_statistics = NULL,
  model_data_constraints = NULL,
  bias = NULL,
  explainability = NULL,
  bias_pre_training = NULL,
  bias_post_training = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_statistics}}{(MetricsSource): A metric source object that represents
model statistics (default: None).}

\item{\code{model_constraints}}{(MetricsSource): A metric source object that represents
model constraints (default: None).}

\item{\code{model_data_statistics}}{(MetricsSource): A metric source object that represents
model data statistics (default: None).}

\item{\code{model_data_constraints}}{(MetricsSource): A metric source object that represents
model data constraints (default: None).}

\item{\code{bias}}{(MetricsSource): A metric source object that represents bias report
(default: None).}

\item{\code{explainability}}{(MetricsSource): A metric source object that represents
explainability report (default: None).}

\item{\code{bias_pre_training}}{(MetricsSource): A metric source object that represents
Pre-training report (default: None).}

\item{\code{bias_post_training}}{(MetricsSource): A metric source object that represents
Post-training report (default: None).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_request_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_request_list}{}}}
\subsection{Method \code{to_request_list()}}{
Generates a request dictionary using the parameters provided to the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMetrics$to_request_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\if{latex}{\out{\hypertarget{method-format}{}}}
\subsection{Method \code{format()}}{
format class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMetrics$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
