% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ModelPackage}
\alias{ModelPackage}
\title{ModelPackage class}
\description{
A SageMaker ``Model`` that can be deployed to an ``Endpoint``.
}
\section{Super classes}{
\code{\link[sagemaker.mlcore:ModelBase]{sagemaker.mlcore::ModelBase}} -> \code{\link[sagemaker.mlcore:Model]{sagemaker.mlcore::Model}} -> \code{ModelPackage}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelPackage$new()}}
\item \href{#method-enable_network_isolation}{\code{ModelPackage$enable_network_isolation()}}
\item \href{#method-.create_sagemaker_model}{\code{ModelPackage$.create_sagemaker_model()}}
\item \href{#method-clone}{\code{ModelPackage$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="ModelBase" data-id="format">}\href{../../sagemaker.mlcore/html/ModelBase.html#method-format}{\code{sagemaker.mlcore::ModelBase$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="ModelBase" data-id="help">}\href{../../sagemaker.mlcore/html/ModelBase.html#method-help}{\code{sagemaker.mlcore::ModelBase$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Model" data-id="check_neo_region">}\href{../../sagemaker.mlcore/html/Model.html#method-check_neo_region}{\code{sagemaker.mlcore::Model$check_neo_region()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Model" data-id="compile">}\href{../../sagemaker.mlcore/html/Model.html#method-compile}{\code{sagemaker.mlcore::Model$compile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Model" data-id="delete_model">}\href{../../sagemaker.mlcore/html/Model.html#method-delete_model}{\code{sagemaker.mlcore::Model$delete_model()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Model" data-id="deploy">}\href{../../sagemaker.mlcore/html/Model.html#method-deploy}{\code{sagemaker.mlcore::Model$deploy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Model" data-id="package_for_edge">}\href{../../sagemaker.mlcore/html/Model.html#method-package_for_edge}{\code{sagemaker.mlcore::Model$package_for_edge()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Model" data-id="prepare_container_def">}\href{../../sagemaker.mlcore/html/Model.html#method-prepare_container_def}{\code{sagemaker.mlcore::Model$prepare_container_def()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Model" data-id="register">}\href{../../sagemaker.mlcore/html/Model.html#method-register}{\code{sagemaker.mlcore::Model$register()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="Model" data-id="transformer">}\href{../../sagemaker.mlcore/html/Model.html#method-transformer}{\code{sagemaker.mlcore::Model$transformer()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a SageMaker ModelPackage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackage$new(
  role,
  model_data = NULL,
  algorithm_arn = NULL,
  model_package_arn = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{(str): An AWS IAM role (either name or full ARN). The Amazon
SageMaker training jobs and APIs that create Amazon SageMaker
endpoints use this role to access training data and model
artifacts. After the endpoint is created, the inference code
might use the IAM role, if it needs to access an AWS resource.}

\item{\code{model_data}}{(str): The S3 location of a SageMaker model data
``.tar.gz`` file. Must be provided if algorithm_arn is provided.}

\item{\code{algorithm_arn}}{(str): algorithm arn used to train the model, can be
just the name if your account owns the algorithm. Must also
provide ``model_data``.}

\item{\code{model_package_arn}}{(str): An existing SageMaker Model Package arn,
can be just the name if your account owns the Model Package.
``model_data`` is not required.}

\item{\code{...}}{: Additional kwargs passed to the Model constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-enable_network_isolation"></a>}}
\if{latex}{\out{\hypertarget{method-enable_network_isolation}{}}}
\subsection{Method \code{enable_network_isolation()}}{
Whether to enable network isolation when creating a model out of this
             ModelPackage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackage$enable_network_isolation()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
bool: If network isolation should be enabled or not.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.create_sagemaker_model"></a>}}
\if{latex}{\out{\hypertarget{method-.create_sagemaker_model}{}}}
\subsection{Method \code{.create_sagemaker_model()}}{
Create a SageMaker Model Entity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackage$.create_sagemaker_model(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{: Positional arguments coming from the caller. This class does not require
any so they are ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPackage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
