% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predict.Predictor}
\alias{predict.Predictor}
\title{S3 method that wraps Predictor Class}
\usage{
\method{predict}{Predictor}(object, newdata, serializer = NULL, deserializer = NULL, ...)
}
\arguments{
\item{object}{a sagemaker model}

\item{newdata}{data for model to predict}

\item{serializer}{method class to serializer data to sagemaker model. Requires to be
a class inherited from \link{BaseSerializer}.}

\item{deserializer}{method class to deserializer return data streams from sagemaker model.
Requires to be a class inherited from \link{BaseDeserializer}.}

\item{...}{arguments passed to ``Predictor$predict``}
}
\description{
Predicted values returned from endpoint
}
