% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{PredictorBase}
\alias{PredictorBase}
\title{PredictorBase Class}
\description{
An object that encapsulates a deployed model
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{content_type}}{The MIME type of the data sent to the inference server.}

\item{\code{accept}}{The content type(s) that are expected from the inference server.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-predict}{\code{PredictorBase$predict()}}
\item \href{#method-delete_predictor}{\code{PredictorBase$delete_predictor()}}
\item \href{#method-clone}{\code{PredictorBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Perform inference on the provided data and return a prediction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorBase$predict(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{: not implemented}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_predictor"></a>}}
\if{latex}{\out{\hypertarget{method-delete_predictor}{}}}
\subsection{Method \code{delete_predictor()}}{
Destroy resources associated with this predictor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorBase$delete_predictor(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{: not implemented}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
