% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_common.R
\name{read_records_io}
\alias{read_records_io}
\title{Eagerly read a collection of amazon Record protobuf objects from raw object}
\usage{
read_records_io(obj)
}
\arguments{
\item{obj}{(raw): raw object}
}
\description{
Eagerly read a collection of amazon Record protobuf objects from raw object
}
