% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{sagemaker.mlcore-package}
\alias{sagemaker.mlcore}
\alias{sagemaker.mlcore-package}
\title{r6 sagemaker: this is just a placeholder}
\description{
`sagemaker` machine learning core classes and methods.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.r.jones@gmail.com}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
