% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\name{SparseMatrixSerializer}
\alias{SparseMatrixSerializer}
\title{SparseMatrixSerializer Class}
\description{
Serialize a sparse matrix to a buffer using the .npz format.
}
\seealso{
Other serializer: 
\code{\link{BaseDeserializer}},
\code{\link{BaseSerializer}},
\code{\link{BytesDeserializer}},
\code{\link{CSVDeserializer}},
\code{\link{CSVSerializer}},
\code{\link{DataTableDeserializer}},
\code{\link{IdentitySerializer}},
\code{\link{JSONDeserializer}},
\code{\link{JSONLinesDeserializer}},
\code{\link{JSONLinesSerializer}},
\code{\link{JSONSerializer}},
\code{\link{LibSVMSerializer}},
\code{\link{NumpyDeserializer}},
\code{\link{NumpySerializer}},
\code{\link{SimpleBaseDeserializer}},
\code{\link{SimpleBaseSerializer}},
\code{\link{StringDeserializer}},
\code{\link{TibbleDeserializer}}
}
\concept{serializer}
\section{Super classes}{
\code{\link[sagemaker.mlcore:BaseSerializer]{sagemaker.mlcore::BaseSerializer}} -> \code{\link[sagemaker.mlcore:SimpleBaseSerializer]{sagemaker.mlcore::SimpleBaseSerializer}} -> \code{SparseMatrixSerializer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scipy}}{Python scipy package}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SparseMatrixSerializer$new()}}
\item \href{#method-serialize}{\code{SparseMatrixSerializer$serialize()}}
\item \href{#method-clone}{\code{SparseMatrixSerializer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sagemaker.mlcore" data-topic="BaseSerializer" data-id="format">}\href{../../sagemaker.mlcore/html/BaseSerializer.html#method-format}{\code{sagemaker.mlcore::BaseSerializer$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a ``SparseMatrixSerializer`` instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SparseMatrixSerializer$new(content_type = "application/x-npz")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{content_type}}{(str): The MIME type to signal to the inference endpoint when sending
request data (default: "application/x-npz").}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
Serialize a sparse matrix to a buffer using the .npz format.
             Sparse matrices can be in the ``csc``, ``csr``, ``bsr``, ``dia`` or
             ``coo`` formats.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SparseMatrixSerializer$serialize(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(sparseMatrix): The sparse matrix to serialize.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
raw: A buffer containing the serialized sparse matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SparseMatrixSerializer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
