% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_compiler_config.R
\name{TrainingCompilerConfig}
\alias{TrainingCompilerConfig}
\title{The configuration class for accelerating SageMaker training jobs through compilation.}
\description{
SageMaker Training Compiler speeds up training by optimizing the model execution graph.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{DEBUG_PATH}}{Placeholder}

\item{\code{SUPPORTED_INSTANCE_CLASS_PREFIXES}}{Placeholder}

\item{\code{HP_ENABLE_COMPILER}}{Placeholder}

\item{\code{HP_ENABLE_DEBUG}}{Placeholder}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TrainingCompilerConfig$new()}}
\item \href{#method-disclaimers_and_warnings}{\code{TrainingCompilerConfig$disclaimers_and_warnings()}}
\item \href{#method-.to_hyperparameter_list}{\code{TrainingCompilerConfig$.to_hyperparameter_list()}}
\item \href{#method-validate}{\code{TrainingCompilerConfig$validate()}}
\item \href{#method-clone}{\code{TrainingCompilerConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
This class initializes a ``TrainingCompilerConfig`` instance.
             Pass the output of it to the ``compiler_config``
             parameter of the :class:`~sagemaker.huggingface.HuggingFace`
             class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingCompilerConfig$new(enabled = TRUE, debug = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{enabled}}{(bool): Optional. Switch to enable SageMaker Training Compiler.
The default is \code{True}.}

\item{\code{debug}}{(bool): Optional. Whether to dump detailed logs for debugging.
This comes with a potential performance slowdown.
The default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-disclaimers_and_warnings"></a>}}
\if{latex}{\out{\hypertarget{method-disclaimers_and_warnings}{}}}
\subsection{Method \code{disclaimers_and_warnings()}}{
Disclaimers and warnings.
             Logs disclaimers and warnings about the
             requested configuration of SageMaker Training Compiler.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingCompilerConfig$disclaimers_and_warnings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.to_hyperparameter_list"></a>}}
\if{latex}{\out{\hypertarget{method-.to_hyperparameter_list}{}}}
\subsection{Method \code{.to_hyperparameter_list()}}{
Converts configuration object into hyperparameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingCompilerConfig$.to_hyperparameter_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(list): A portion of the hyperparameters passed to the training job as a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
Checks if SageMaker Training Compiler is configured correctly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingCompilerConfig$validate(image_uri, instance_type, distribution)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image_uri}}{(str): A string of a Docker image URI that's specified
to :class:`~sagemaker.huggingface.HuggingFace`.
If SageMaker Training Compiler is enabled, the HuggingFace estimator
automatically chooses the right image URI. You cannot specify and override
the image URI.}

\item{\code{instance_type}}{(str): A string of the training instance type that's specified
to :class:`~sagemaker.huggingface.HuggingFace`.
The `validate` classmethod raises error
if an instance type not in the ``SUPPORTED_INSTANCE_CLASS_PREFIXES`` list
or ``local`` is passed to the `instance_type` parameter.}

\item{\code{distribution}}{(dict): A dictionary of the distributed training option that's specified
to :class:`~sagemaker.huggingface.HuggingFace`.
SageMaker's distributed data parallel and model parallel libraries
are currently not compatible with SageMaker Training Compiler.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrainingCompilerConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
