% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon_validation.R
\name{Validation}
\alias{Validation}
\title{Helper class}
\description{
Check expression
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gt}{\code{Validation$gt()}}
\item \href{#method-ge}{\code{Validation$ge()}}
\item \href{#method-lt}{\code{Validation$lt()}}
\item \href{#method-le}{\code{Validation$le()}}
\item \href{#method-isin}{\code{Validation$isin()}}
\item \href{#method-istype}{\code{Validation$istype()}}
\item \href{#method-clone}{\code{Validation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gt"></a>}}
\if{latex}{\out{\hypertarget{method-gt}{}}}
\subsection{Method \code{gt()}}{
greater than
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validation$gt(minimum)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{minimum}}{(object): placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ge"></a>}}
\if{latex}{\out{\hypertarget{method-ge}{}}}
\subsection{Method \code{ge()}}{
greater equals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validation$ge(minimum)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{minimum}}{(object): placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lt"></a>}}
\if{latex}{\out{\hypertarget{method-lt}{}}}
\subsection{Method \code{lt()}}{
less than
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validation$lt(maximum)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maximum}}{(object): placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-le"></a>}}
\if{latex}{\out{\hypertarget{method-le}{}}}
\subsection{Method \code{le()}}{
less equals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validation$le(maximum)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{maximum}}{(object): placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isin"></a>}}
\if{latex}{\out{\hypertarget{method-isin}{}}}
\subsection{Method \code{isin()}}{
is in
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validation$isin(expected)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{(object): placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-istype"></a>}}
\if{latex}{\out{\hypertarget{method-istype}{}}}
\subsection{Method \code{istype()}}{
is data type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validation$istype(expected)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{(object): placeholder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
