% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuner.R
\docType{data}
\name{WarmStartTypes}
\alias{WarmStartTypes}
\title{WarmStartTypes Class}
\format{
An object of class \code{WarmStartTypes} (inherits from \code{Enum}, \code{environment}) of length 2.
}
\usage{
WarmStartTypes
}
\description{
Warm Start Configuration type. There can be two types of warm start jobs:
             * IdenticalDataAndAlgorithm: Type of warm start that allows users to reuse
             training results from existing tuning jobs that have the same algorithm code
             and datasets.
             * TransferLearning: Type of warm start that allows users to
             reuse training results from existing tuning jobs that have similar algorithm
             code and datasets.
}
\keyword{datasets}
